requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==1.0.1
astor==0.8.1
astroid==2.3.1
astropy==3.2.2
astunparse==1.6.3
atari-py==0.2.6
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
basemap @ https://github.com/matplotlib/basemap/archive/master.zip
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
boto3==1.14.11
botocore==1.17.11
Bottleneck==1.2.1
branca==0.4.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.12.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.1
clyent==1.2.2
colorama==0.4.1
conda==4.7.12
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0
cryptography==2.7
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decopatch==1.4.7
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.2
distributed==2.5.2
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
findspark==1.3.0
Flask==1.1.2
folium==0.11.0
fsspec==0.5.2
future==0.18.2
gast==0.3.3
gensim==3.8.3
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
google==2.0.3
google-auth==1.19.0
google-auth-oauthlib==0.4.1
google-colab==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.30.0
gunicorn==20.0.4
gym==0.14.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
idna==2.10
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==1.7.0
ipykernel==5.1.4
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.0.0
jupyter-core==4.6.3
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras==2.1.0
Keras-Applications==1.0.8
keras-gym==0.2.17
Keras-Preprocessing==1.1.2
keyring==18.0.0
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
makefun==1.9.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.1
matplotlib-venn==0.11.5
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.45.1
numexpr==2.7.0
numpy==1.19.0
numpy-stl==2.11.2
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.32
openpyxl==3.0.0
opt-einsum==3.2.1
packaging==19.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
plotly==4.8.1
pluggy==0.13.0
ply==3.11
portpicker==1.2.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==1.0.18
protobuf==3.12.2
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pydataset==0.2.0
pyflakes==2.1.1
pygame==1.9.6
pyglet==1.3.2
Pygments==2.4.2
pylint==2.4.2
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyproj==2.6.1.post1
pyrsistent==0.15.4
pyshp==2.1.0
PySocks==1.7.1
pyspark==2.4.5
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-cases==1.13.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.0
python-dotenv==0.14.0
python-utils==2.4.0
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.14.0
rsa==4.6
ruamel-yaml==0.15.46
s3transfer==0.3.3
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0
smart-open==2.0.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
statsmodels==0.10.1
sympy==1.4
tables==3.5.2
tblib==1.4.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-datasets==3.2.0
tensorflow-estimator==2.2.0
tensorflow-metadata==0.22.2
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
torch==1.5.1
torchaudio==0.5.1
torchvision==0.5.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
unicodecsv==0.14.1
urllib3==1.25.9
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.7.0
wrapt==1.12.1
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
zict==1.0.0
zipp==3.1.0
